Eliminate a single-use TractionBC module and fold its functionality into Mechanics
